x86: normalize types used with _mm_set?_epi64x
w/clang -fsanitize=integer fixes warnings of the form:
implicit conversion from type 'uint64_t' (aka 'unsigned long') of value
15120695205995520509 (64-bit, unsigned) to type 'long long' changed the
value to -3326048867714031107 (64-bit, signed)

Bug: aomedia:3136
Bug: b/229626362

+++ aom_dsp/x86/synonyms.h
@@ -70,7 +70,7 @@ static INLINE __m128i xx_set_64_from_32i(int32_t e1, int32_t e0) {
 #if defined(_MSC_VER) && _MSC_VER < 1900
   return _mm_set_epi32(0, e1, 0, e0);
 #else
-  return _mm_set_epi64x((uint32_t)e1, (uint32_t)e0);
+  return _mm_set_epi64x(e1, e0);
 #endif
 }
 
@@ -81,7 +81,7 @@ static INLINE __m128i xx_set1_64_from_32i(int32_t a) {
 #if defined(_MSC_VER) && _MSC_VER < 1900
   return _mm_set_epi32(0, a, 0, a);
 #else
-  return _mm_set1_epi64x((uint32_t)a);
+  return _mm_set1_epi64x(a);
 #endif
 }
